Backend/requirements.txt
flask
flask-restful
flask-sqlalchemy
python-dotenv
flask-marshmallow
marshmallow-sqlalchemy
passlib
flask-jwt-extended
flask-cors
Flask-Migrate